+++ Engine_SOURCE/Animation.cpp
@@ -53,7 +53,7 @@ namespace dru
 				if (mSpriteSheet.size() <= mIndex)
 				{
 					mbComplete = true;
-					mIndex = mSpriteSheet.size() - 1;
+					mIndex = static_cast<int>(mSpriteSheet.size() - 1);
 
 				}
