+++ Engine_SOURCE/Material.cpp
@@ -58,7 +58,7 @@ namespace dru::graphics
 	}
 	void CMaterial::Bind()
 	{
-		for (size_t i = 0; i < (UINT)eTextureSlot::End; i++)
+		for (UINT i = 0; i < (UINT)eTextureSlot::End; i++)
         {
             if (mTexture[i] == nullptr)
                 continue;
